modified scuff-transmission to handle calculation of transmission amplitudes from surface currents

+++ src/applications/scuff-scatter/SIPFT.cc
@@ -548,7 +548,7 @@ void GetSIPFTMatrixEntries(RWGSurface *S, int neA, int neB,
 /***************************************************************/
 /***************************************************************/
 void GetSIPFTMatrices(RWGGeometry *G, int WhichSurface,
-                      RWGSurface *BS, int R, int NumPoints,
+                      RWGSurface *BS, double R, int NumPoints,
                       cdouble Omega, bool NeedMatrix[NUMSIPFT],
                       HMatrix *MSIPFT[NUMSIPFT])
 {
